Revert "Change type of cost & priority to double"
This reverts commit eb2107e4cc08425b9e0cbc1df586d26a7cf98a96.

+++ Wurst Client/src/tk/wurst_client/ai/PathPoint.java
@@ -16,11 +16,11 @@ public class PathPoint
 {
 	private BlockPos pos;
 	private PathPoint previous;
-	private double priority;
-	private double movementCost;
+	private int priority;
+	private int movementCost;
 	
-	public PathPoint(BlockPos pos, PathPoint previous, double movementCost,
-		double priority)
+	public PathPoint(BlockPos pos, PathPoint previous, int movementCost,
+		int priority)
 	{
 		this.pos = pos;
 		this.previous = previous;
@@ -70,12 +70,12 @@ public class PathPoint
 		return previous;
 	}
 	
-	public double getPriority()
+	public int getPriority()
 	{
 		return priority;
 	}
 	
-	public double getMovementCost()
+	public int getMovementCost()
 	{
 		return movementCost;
 	}
